+++ syndicate/core/resources/api_gateway_resource.py
@@ -22,7 +22,9 @@ from botocore.exceptions import ClientError
 from syndicate.commons import deep_get
 from syndicate.commons.log_helper import get_logger
 from syndicate.core.constants import (
-    SOURCE_ARN_DEEP_KEY, SECURITY_SCHEMAS_DEEP_KEY
+    SOURCE_ARN_DEEP_KEY, SECURITY_SCHEMAS_DEEP_KEY,
+    API_GW_DEFAULT_THROTTLING_RATE_LIMIT,
+    API_GW_DEFAULT_THROTTLING_BURST_LIMIT
 )
 from syndicate.core.helper import unpack_kwargs
 from syndicate.core.resources.base_resource import BaseResource
@@ -68,6 +70,7 @@ _REQUEST_VALIDATORS = {
     }
 }
 
+_DISABLE_THROTTLING_VALUE = -1
 
 class ApiGatewayResource(BaseResource):
 
@@ -213,23 +216,37 @@ class ApiGatewayResource(BaseResource):
         parameter = parameter[:index] + '~1' + parameter[index + 1:]
         return self._escape_path(parameter)
 
-    def configure_cache(self, api_id, stage_name, api_resources):
+    def configure_resources(self, api_id, stage_name, api_resources):
         for resource_path, resource_meta in api_resources.items():
             for method_name, method_meta in resource_meta.items():
                 if method_name in SUPPORTED_METHODS:
                     cache_configuration = method_meta.get(
                         'cache_configuration')
-                    if not cache_configuration:
-                        continue
+                    throttling_configuration = method_meta.get(
+                        'throttling_configuration')
                     cache_ttl_setting = cache_configuration.get(
-                        'cache_ttl_sec')
+                        'cache_ttl_sec') if cache_configuration else None
                     encrypt_cache_data = cache_configuration.get(
-                        'encrypt_cache_data')
+                        'encrypt_cache_data') if cache_configuration else None
+                    throttling_enabled = throttling_configuration.get(
+                        'throttling_enabled') if throttling_configuration \
+                        else None
+                    throttling_rate_limit = throttling_configuration.get(
+                        'throttling_rate_limit',
+                        API_GW_DEFAULT_THROTTLING_RATE_LIMIT) if (
+                        throttling_configuration and throttling_enabled) else \
+                        _DISABLE_THROTTLING_VALUE
+                    throttling_burst_limit = throttling_configuration.get(
+                        'throttling_burst_limit',
+                        API_GW_DEFAULT_THROTTLING_BURST_LIMIT) if (
+                        throttling_configuration and throttling_enabled) else \
+                        _DISABLE_THROTTLING_VALUE
+                    patch_operations = []
+                    escaped_resource = self._escape_path(resource_path)
                     if cache_ttl_setting is not None:
                         _LOG.info(
                             'Configuring cache for {0}; TTL: {1}'.format(
                                 resource_path, cache_ttl_setting))
-                        escaped_resource = self._escape_path(resource_path)
                         patch_operations = [
                             {
                                 'op': 'replace',
@@ -255,14 +272,37 @@ class ApiGatewayResource(BaseResource):
                                 'value': 'true' if bool(
                                     encrypt_cache_data) else 'false'
                             })
-                        self.connection.update_configuration(
-                            rest_api_id=api_id,
-                            stage_name=stage_name,
-                            patch_operations=patch_operations
-                        )
+                    if throttling_enabled:
                         _LOG.info(
-                            'Cache for {0} was configured'.format(
-                                resource_path))
+                            'Configuring throttling for {0}; rateLimit: {1}; '
+                            'burstLimit: {2}'.format(
+                                resource_path, throttling_rate_limit,
+                                throttling_burst_limit))
+                    else:
+                        _LOG.info('Throttling for {0} disabled.'.format(
+                            resource_path))
+                    patch_operations.append({
+                            'op': 'replace',
+                            'path': '/{0}/{1}/throttling/rateLimit'
+                                    ''.format(escaped_resource,
+                                              method_name),
+                            'value': str(throttling_rate_limit),
+                    })
+                    patch_operations.append({
+                            'op': 'replace',
+                            'path': '/{0}/{1}/throttling/burstLimit'
+                                    ''.format(escaped_resource,
+                                              method_name),
+                            'value': str(throttling_burst_limit),
+                    })
+                    self.connection.update_configuration(
+                        rest_api_id=api_id,
+                        stage_name=stage_name,
+                        patch_operations=patch_operations
+                    )
+                    _LOG.info(
+                        'Resource {0} was configured'.format(
+                            resource_path))
 
     @unpack_kwargs
     def _create_api_gateway_from_meta(self, name, meta):
@@ -440,12 +480,27 @@ class ApiGatewayResource(BaseResource):
                                    cache_cluster_enabled=root_cache_enabled,
                                    cache_cluster_size=str(
                                        cache_size) if cache_size else None)
+        throttling_cluster_configuration = meta.get(
+            'cluster_throttling_configuration')
+        throttling_enabled = throttling_cluster_configuration.get(
+            'throttling_enabled') if throttling_cluster_configuration else None
+        patch_operations = []
+        if not throttling_enabled:
+            patch_operations.append({
+                'op': 'replace',
+                'path': '/*/*/throttling/rateLimit',
+                'value': str(_DISABLE_THROTTLING_VALUE),
+            })
+            patch_operations.append({
+                'op': 'replace',
+                'path': '/*/*/throttling/burstLimit',
+                'value': str(_DISABLE_THROTTLING_VALUE),
+            })
         # configure caching
         if root_cache_enabled:
             _LOG.debug('Cluster cache configuration found:{0}'.format(
                 cache_cluster_configuration))
             # set default ttl for root endpoint
-            patch_operations = []
             cluster_cache_ttl_sec = cache_cluster_configuration.get(
                 'cache_ttl_sec')
             encrypt_cache_data = cache_cluster_configuration.get(
@@ -462,14 +517,30 @@ class ApiGatewayResource(BaseResource):
                     'path': '/*/*/caching/dataEncrypted',
                     'value': 'true' if bool(encrypt_cache_data) else 'false'
                 })
-            if patch_operations:
-                self.connection.update_configuration(
+        # configure throttling
+        if throttling_enabled:
+            throttling_rate_limit = throttling_cluster_configuration.get(
+                'throttling_rate_limit', API_GW_DEFAULT_THROTTLING_RATE_LIMIT)
+            throttling_burst_limit = throttling_cluster_configuration.get(
+                'throttling_burst_limit', API_GW_DEFAULT_THROTTLING_BURST_LIMIT)
+            patch_operations.append({
+                'op': 'replace',
+                'path': '/*/*/throttling/rateLimit',
+                'value': str(throttling_rate_limit),
+            })
+            patch_operations.append({
+                'op': 'replace',
+                'path': '/*/*/throttling/burstLimit',
+                'value': str(throttling_burst_limit),
+            })
+        if any([root_cache_enabled, throttling_enabled]):
+            self.connection.update_configuration(
                     rest_api_id=api_id,
                     stage_name=deploy_stage,
                     patch_operations=patch_operations
                 )
-            # customize cache settings for endpoints
-            self.configure_cache(api_id, deploy_stage, api_resources)
+        # customize settings for endpoints
+        self.configure_resources(api_id, deploy_stage, api_resources)
 
     def __prepare_api_resources_args(
             self, api_id, api_resources, api_resp=None,