+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [1.13.2] - 2024-02-05
+# [1.14.1] - 2024-02-05
 - Improve the CloudWatch log groups `POSSIBLE_RETENTION_DAYS` constant to
 support all values.
 - Update and fix retry decorator. If the maximum number of retries is reached,
@@ -12,6 +12,12 @@ an `Exception` is thrown with a detailed description of the function that failed
 and the reason.
 - Refactor the Lambda SQS Trigger Creation process to check for existing event
 source mapping, and update or create as needed.
+
+# [1.14.0] - 2024-02-05
+- Implemented API GateWay `Throttling` settings management
+
+# [1.13.2] - 2024-02-05
+- Added CloudWatch logs expiration management support to the Syndicate Java plugin 1.11.0
 
 # [1.13.1] - 2024-02-01
 - Lambda parameter `architecture` changed to `architectures`, and the value of the key changed to a list of string